moving k6 extensions to the extensions subfolder
