+++ extensions/k6/Makefile
@@ -1,0 +1,3 @@
+.PHONY: build
+build:
+	xk6 build --with github.com/kubeshop/tracetest/extensions/k6=.
