+++ extensions/k6/examples/async/test-from-definition.js
