+++ extensions/k6/examples/async/test-from-id.js
