+++ extensions/k6/examples/sync/test-from-definition.js
