+++ extensions/k6/modules/httpClient/enhancers.go
@@ -5,8 +5,8 @@ import (
 	"time"
 
 	"github.com/dop251/goja"
-	"github.com/kubeshop/tracetest/k6-extension/models"
-	"github.com/kubeshop/tracetest/k6-extension/utils"
+	"github.com/kubeshop/tracetest/extensions/k6/models"
+	"github.com/kubeshop/tracetest/extensions/k6/utils"
 	"go.k6.io/k6/lib"
 	"go.k6.io/k6/metrics"
 )
