+++ extensions/k6/modules/httpClient/options.go
@@ -6,8 +6,8 @@ import (
 	"strings"
 
 	"github.com/dop251/goja"
-	"github.com/kubeshop/tracetest/k6-extension/models"
-	"github.com/kubeshop/tracetest/k6-extension/utils"
+	"github.com/kubeshop/tracetest/extensions/k6/models"
+	"github.com/kubeshop/tracetest/extensions/k6/utils"
 	"go.k6.io/k6/js/modules"
 	k6HTTP "go.k6.io/k6/js/modules/k6/http"
 )
