+++ extensions/k6/tracetest.go
@@ -1,9 +1,9 @@
 package tracetest
 
 import (
-	"github.com/kubeshop/tracetest/k6-extension/modules/instance"
-	tracetestOutput "github.com/kubeshop/tracetest/k6-extension/modules/output"
-	"github.com/kubeshop/tracetest/k6-extension/modules/tracetest"
+	"github.com/kubeshop/tracetest/extensions/k6/modules/instance"
+	tracetestOutput "github.com/kubeshop/tracetest/extensions/k6/modules/output"
+	"github.com/kubeshop/tracetest/extensions/k6/modules/tracetest"
 	"go.k6.io/k6/js/modules"
 	"go.k6.io/k6/output"
 )
