--- k6-extension/Makefile
@@ -1,3 +1,0 @@
-.PHONY: build
-build:
-	xk6 build --with github.com/kubeshop/tracetest/k6-extension=.
